Refactored directory names
