+++ frontend/src/components/6.Swiping/SwipingPage.css
