+++ frontend/src/components/7.Result/DataVisual.js
