+++ frontend/src/components/7.Result/ResultPage.js
